services/crm_data_entry_agent/requirements.txt
Flask>=2.3.0
hubspot-api-client
